Fiz extra fields
